tunneling: Simplify path MTU discovery logic.
Previously, if path MTU discovery was disabled we would use the
tunnel MTU instead of the underlying route's path MTU but otherwise
still do PMTUD.  This doesn't make much sense because turning off
PMTUD really means to not check the size of the encapsulated packets
at all.  This removes the disconnect and simplifies the logic.

Signed-off-by: Jesse Gross <jesse@nicira.com>
Acked-by: Ben Pfaff <blp@nicira.com>