vswitchd: Use wdp instead of xfif interface from bridge code.
Part of the point of the wdp branch is to use the high-level wdp interface
where it makes sense.  But until now this hasn't been extended to the
bridge code.  This commit fixes that.

There are still some references to xflow inside this file that need to be
looked at.