ovsdb: Don't count self-references in ovsdb_row's n_refs member.
The comment on the n_refs member of struct ovsdb_row says that it does not
count references from a row to itself, but the code didn't implement this
properly.  This commit makes the code consistent with the comment.

This does not actually affect any existing OVSDB behavior, because a row's
reference count currently affects only whether it may be deleted, and
references from a row to itself disappear when the row is deleted.  But
an upcoming commit will add new uses for a row's reference count, so at
that point it becomes important.