lacp: Fix misleading prototype for lacp_configure().
Only the first 6 bytes (ETH_ADDR_LEN) of the 'sys_id' argument are used,
but the prototype declared it as an array of 8 bytes.  This has no effect
on the generated code--the declared size of an array parameter is
irrelevant--but it is misleading.

Also, add 'const' since the array is not modified.