docs: Suppress "warning: macro `DD' not defined" warnings from man.
deamon.man allows the file that is including it to include extra text in
the description of --detach by defining a macro named DD.  Only some of the
manpages that included it did this (only those manpages that needed extra
text there).  But it's better to be quiet in "man --warnings", so this
defines DD to an empty value in the other manpages that include daemon.man.

Reported by lintian.